refactor: update DTO namespaces
Updated DTO namespaces and file structure
