+++ app/Application/DTOs/Inputs/CreateUserInput.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Application\DTOs\Input;
+namespace App\Application\DTOs\Inputs;
 
 use InvalidArgumentException;
